Moves routes/completer to their own modules
